requirements.txt
Django==3.2.7
django-ckeditor==6.5.1
django-cors-headers==3.7.0
django-currencies==0.10.1
django-filter==23.2
django-js-asset==2.0.0
django-shopping-cart==0.1
django-storages==1.11.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
duplicity==0.8.22
exceptiongroup==1.1.3
executing==1.2.0
future==0.18.2
gunicorn==20.1.0
httplib2==0.20.4
ipykernel==6.23.2
numpy==1.24.3
Pillow==10.0.0
pip-chill==1.0.3
psycopg2-binary==2.9.6
pycodestyle==2.7.0

